Remove obsolete workaround for ESP8266 in compilation test CI workflow (#38)
The ESP8266 platform has a tool dependency on python3, but assumes that tool is already installed in a particular location on Linux. At the time this workaround was added, the Docker container used by the arduino/actions/libraries/compile-examples action did not provide this tool, so a workaround was needed to install it in the container. Since that time, the action has developed a dependency on Python, making this workaround unnecessary. The switch from the action being a Bash script to a Python script caused the now-unnecessary workaround to break the workflow.